off_t for file size > 4G

+++ lego.ev3/system/drivers/adcd/main.c
@@ -33,7 +33,7 @@ uint16_t _adc_buf[MAX_CH];
 uint8_t	 _ch = 0;
 
 static int adcd_read(vdevice_t* dev, int fd, int from_pid, fsinfo_t* node, 
-        void* buf, int size, int offset, void* p) {
+        void* buf, int size, off_t offset, void* p) {
     (void)dev;
     (void)fd;
     (void)from_pid;
