+++ raspi5/system/drivers/fand/fand.c
@@ -335,7 +335,7 @@ static char* fan_cmd(vdevice_t* dev, int from_pid, int argc, char** argv, void* 
 }
 
 static int fan_read(vdevice_t* dev, int fd, int from_pid, fsinfo_t* info,
-        void* buf, int size, int offset, void* p) {
+        void* buf, int size, off_t offset, void* p) {
     (void)dev; (void)fd; (void)from_pid; (void)info; (void)p;
 
     if (size <= 0 || offset != 0)
@@ -346,7 +346,7 @@ static int fan_read(vdevice_t* dev, int fd, int from_pid, fsinfo_t* info,
 }
 
 static int fan_write(vdevice_t* dev, int fd, int from_pid, fsinfo_t* info,
-        const void* buf, int size, int offset, void* p) {
+        const void* buf, int size, off_t offset, void* p) {
     (void)dev; (void)fd; (void)from_pid; (void)info; (void)offset; (void)p;
 
     if (size <= 0)
