+++ raspix/3rd/others/drivers/spi2uartd/uartd.c
@@ -78,7 +78,7 @@ static inline uint8_t sc16is750_other_channel(uint8_t channel) {
 }
 
 static int uart_read(vdevice_t* dev, int fd, int from_pid, fsinfo_t* node,
-                void* buf, int size, int offset, void* p) {
+                void* buf, int size, off_t offset, void* p) {
         (void)dev;
         (void)fd;
         (void)from_pid;
@@ -97,7 +97,7 @@ static int uart_read(vdevice_t* dev, int fd, int from_pid, fsinfo_t* node,
 }
 
 static int uart_write(vdevice_t* dev, int fd, int from_pid, fsinfo_t* node,
-                const void* buf, int size, int offset, void* p) {
+                const void* buf, int size, off_t offset, void* p) {
         (void)dev;
         (void)fd;
         (void)node;
