+++ x2lite.rk3128/system/drivers/gpio_joystickd/joystickd.c
@@ -154,7 +154,7 @@ static int rockchip_gpio_get_value(struct gpio_pins* pins)
 }
 
 static int joystick_read(vdevice_t* dev, int fd, int from_pid, fsinfo_t* node,
-        void* buf, int size, int offset, void* p) {
+        void* buf, int size, off_t offset, void* p) {
     (void)dev;
     (void)fd;
     (void)from_pid;
