+++ x86/system/drivers/ps2keybd/keybd.c
@@ -229,7 +229,7 @@ static void keybd_interrupt_handle(uint32_t interrupt, ewokos_addr_t data) {
 }
 
 static int keybd_read(vdevice_t* dev, int fd, int from_pid, fsinfo_t* info,
-        void* buf, int size, int offset, void* p) {
+        void* buf, int size, off_t offset, void* p) {
     (void)dev;
     (void)fd;
     (void)from_pid;
